add Render.WriteImage and Render.ReadImage
